refactor(auth): align with NextAuth.js v5 best practices
- Centralizes the NextAuth `auth` function in `src/lib/auth-server.ts`.
- Updates all API routes to import `auth` from the new central location.
- Removes the redundant `auth` export from the `[...nextauth]` route handler.
- Fixes the `getServerSession` build error by replacing it with the new `auth` function.
