+++ src/app/api/analytics/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-import { auth } from "@/app/api/auth/[...nextauth]/route";
+import { auth } from "@/lib/auth-server";
 import { prisma } from "@/lib/prisma";
 import { POStatus } from "@/types/po";
