+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,8 +1,3 @@
 // src/app/api/auth/[...nextauth]/route.ts
-import NextAuth from "next-auth";
-import { authOptions } from "@/lib/auth-options";
-
-// Use the shared authOptions
-const { handlers: { GET, POST }, auth } = NextAuth(authOptions);
-
-export { GET, POST, auth };
+import { handlers } from "@/lib/auth-server";
+export const { GET, POST } = handlers;
