+++ src/app/api/cr/po/route.ts
@@ -1,6 +1,6 @@
 // src/app/api/cr/po/route.ts
 import { NextRequest, NextResponse } from "next/server";
-import { auth } from "@/app/api/auth/[...nextauth]/route";
+import { auth } from "@/lib/auth-server";
 import { getPOsForCR } from "@/lib/cr";
 
 export async function GET() {
