+++ src/app/api/notifications/[id]/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-import { auth } from "@/app/api/auth/[...nextauth]/route";
+import { auth } from "@/lib/auth-server";
 import { markNotificationAsRead } from "@/lib/notification";
 
 interface RouteParams {
