+++ src/app/api/notifications/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-import { auth } from "@/app/api/auth/[...nextauth]/route";
+import { auth } from "@/lib/auth-server";
 import { getNotifications } from "@/lib/notification";
 
 export async function GET(request: NextRequest) {
