+++ src/app/api/po/iom/route.ts
@@ -1,6 +1,6 @@
 // src/app/api/po/iom/route.ts
 import { NextRequest, NextResponse } from "next/server";
-import { auth } from "@/app/api/auth/[...nextauth]/route";
+import { auth } from "@/lib/auth-server";
 import { prisma } from "@/lib/prisma";
 
 export async function GET() {
