+++ src/app/api/vendors/route.ts
@@ -1,6 +1,6 @@
 // src/app/api/vendors/route.ts
 import { NextRequest, NextResponse } from "next/server";
-import { auth } from "@/app/api/auth/[...nextauth]/route";
+import { auth } from "@/lib/auth-server";
 import { getVendors, createVendor } from "@/lib/po";
 import { createVendorSchema } from "@/lib/schemas";
